control props patterns completed
