+++ src/compound-component-patterns/data/products.ts
@@ -1,0 +1,16 @@
+import { Product } from "../interfaces";
+
+const products: Product[] = [
+  {
+    id: '1',
+    title: 'Coffee Mug - 001',
+    img: './coffee-mug.png'
+  },
+  {
+    id: '2',
+    title: 'Coffee Mug - 002',
+    img: './coffee-mug2.png'
+  }
+];
+
+export default products;
